manupulacion_de_datos_python/requirements.txt
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
fpdf2==2.7.6
numpy==1.24.3 